Trait migration_rocksdb::Migration

pub trait Migration: 'static {
    fn columns(&self) -> Option<u32>;

    fn version(&self) -> u32;

    fn migrate(
        &mut self, 
        source: Arc<Database>, 
        config: &Config, 
        destination: &mut Database, 
        col: Option<u32>
    ) -> Result<()>;

    fn pre_columns(&self) -> Option<u32> { ... }

    fn alters_existing(&self) -> bool { ... }
}

A generalized migration from the given db to a destination db.

Required methods

fn columns(&self) -> Option<u32>

Number of columns in database after the migration.

fn version(&self) -> u32

Version of the database after the migration.

fn migrate(
&mut self,
source: Arc<Database>,
config: &Config,
destination: &mut Database,
col: Option<u32>
) -> Result<()>

Migrate a source to a destination.

Provided methods

fn pre_columns(&self) -> Option<u32>

Number of columns in the database before the migration.

fn alters_existing(&self) -> bool

Whether this migration alters any existing columns. if not, then column families will simply be added and migrate will never be called.